Implemented mailstatus, that shows me the status of my reddit mail.
If you really are desperate and want to use this: Include it with a
1-minute refresh period. Add a file cookie.txt to the binary dir, and
write your password in the first line. Your username has to be changed in
code. This is because the password solution sucks and I don't want anyone
to use this. Be careful when commiting your project. You don't want to
commit your password.